+++ lib/src/element/score_partwise.dart
@@ -9,6 +9,7 @@ import '../local.dart';
 import 'movement_number.dart';
 import 'partlist/part_list.dart';
 import 'version.dart';
+import 'credit.dart';
 import 'defaults/defaults.dart';
 import 'identification/identification.dart';
 import 'work/work.dart';
@@ -24,7 +25,7 @@ class ScorePartwise extends XmlElement {
   final Identification? identification;
   final Defaults? defaults;
 
-  // TODO final List<Credit> credits; // Zero or more times
+  final List<Credit> credits;
   final PartList partList;
 
   // One or more times
@@ -44,6 +45,7 @@ class ScorePartwise extends XmlElement {
     Work? work;
     Identification? identification;
     Defaults? defaults;
+    final credits = <Credit>[];
     MovementNumber? movementNumber = null;
     MovementTitle? movementTitle = null;
     final partList = PartList.parse(element.getElement(Local.partList)!);
@@ -58,6 +60,9 @@ class ScorePartwise extends XmlElement {
           break;
         case Local.defaults:
           defaults = Defaults.parse(e);
+          break;
+        case Local.credit:
+          credits.add(Credit.parse(e));
           break;
         case Local.movementNumber:
           movementNumber = MovementNumber.parse(e);
@@ -77,6 +82,7 @@ class ScorePartwise extends XmlElement {
       work: work,
       identification: identification,
       defaults: defaults,
+      credits: credits,
       movementNumber: movementNumber,
       movementTitle: movementTitle,
       partList: partList,
@@ -90,6 +96,7 @@ class ScorePartwise extends XmlElement {
     this.work,
     this.identification,
     this.defaults,
+    this.credits = const [],
     this.movementNumber,
     this.movementTitle,
     // this.defaults,
@@ -104,6 +111,7 @@ class ScorePartwise extends XmlElement {
             if (work != null) work,
             if (identification != null) identification,
             if (defaults != null) defaults,
+            ...credits,
             if (movementNumber != null) movementNumber,
             if (movementTitle != null) movementTitle,
             partList,